% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimBM.R
\name{PlotBM}
\alias{PlotBM}
\title{PlotBM}
\usage{
PlotBM(x)
}
\arguments{
\item{x}{2-D Brownian motion.}
}
\value{
None, unless there is any error message.
}
\description{
Plot two-dimensional Brownian motion
}
\examples{
y <- SimBM(n=10000, sigma=1.2)
PlotBM(y)
}

