% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimBM.R
\name{SimBM}
\alias{SimBM}
\title{SimBM}
\usage{
SimBM(n, sigma = 1)
}
\arguments{
\item{n}{Length of Brownian motion.}

\item{sigma}{Standard deviation for error process.}
}
\value{
return the dataset you want
}
\description{
Simulation of two-dimensional Brownian motion
}
\examples{
y <- SimBM(n=10000, sigma=1.2)
}

